feat: 用户职位功能完善

+++ passiflora-server/modules/passiflora-codegen/src/main/resources/templates/service.java.ftl
@@ -1,5 +1,6 @@
 package com.zerolinck.passiflora.${moduleName}.service;
 
+import cn.hutool.core.collection.CollectionUtil;
 import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
@@ -12,6 +13,8 @@ import com.zerolinck.passiflora.model.${moduleName}.entity.${entityClass};
 import com.zerolinck.passiflora.${moduleName}.mapper.${mapperClass};
 import com.zerolinck.passiflora.common.api.ResultCodeEnum;
 import com.zerolinck.passiflora.common.exception.BizException;
+import jakarta.annotation.Nonnull;
+import jakarta.annotation.Nullable;
 import lombok.extern.slf4j.Slf4j;
 import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
@@ -32,17 +35,18 @@ import java.util.concurrent.TimeUnit;
 public class ${serviceClass} extends ServiceImpl<${mapperClass}, ${entityClass}> {
     
     private static final String LOCK_KEY = "passiflora:lock:${entityName}:";
-    
-    public Page<${entityClass}> page(QueryCondition<${entityClass}> condition) {
+
+    @Nonnull
+    public Page<${entityClass}> page(@Nullable QueryCondition<${entityClass}> condition) {
         if (condition == null) {
             condition = new QueryCondition<>();
         }
         return baseMapper.page(condition.page(), condition.searchWrapper(${entityClass}.class), condition.sortWrapper(${entityClass}.class));
     }
 
-    public void add(${entityClass} ${entityName}) {
-        LockUtil.lockAndTransactionalLogic(LOCK_KEY,
-                new LockWrapper<${entityClass}>(),
+    public void add(@Nonnull ${entityClass} ${entityName}) {
+        LockUtil.lock(LOCK_KEY,
+                new LockWrapper<${entityClass}>(), true,
                 () -> {
                     OnlyFieldCheck.checkInsert(baseMapper, ${entityName});
                     baseMapper.insert(${entityName});
@@ -51,9 +55,9 @@ public class ${serviceClass} extends ServiceImpl<${mapperClass}, ${entityClass}>
         );
     }
 
-    public boolean update(${entityClass} ${entityName}) {
-        return (boolean) LockUtil.lockAndTransactionalLogic(LOCK_KEY,
-                new LockWrapper<${entityClass}>(),
+    public boolean update(@Nonnull ${entityClass} ${entityName}) {
+        return (boolean) LockUtil.lock(LOCK_KEY,
+                new LockWrapper<${entityClass}>(), true,
                 () -> {
                     OnlyFieldCheck.checkUpdate(baseMapper, ${entityName});
                     int changeRowCount = baseMapper.updateById(${entityName});
@@ -63,11 +67,15 @@ public class ${serviceClass} extends ServiceImpl<${mapperClass}, ${entityClass}>
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public int deleteByIds(Collection<String> ${table.pkFieldName}s) {
+    public int deleteByIds(@Nullable Collection<String> ${table.pkFieldName}s) {
+        if (CollectionUtil.isEmpty(${table.pkFieldName}s) {
+            return 0;
+        }
         return baseMapper.deleteByIds(${table.pkFieldName}s, CurrentUtil.getCurrentUserId());
     }
 
-    public ${entityClass} detail(String ${table.pkFieldName}) {
+    @Nonnull
+    public ${entityClass} detail(@Nonnull String ${table.pkFieldName}) {
         ${entityClass} ${entityName} = baseMapper.selectById(${table.pkFieldName});
         if (${entityName} == null) {
             throw new BizException("无对应${table.description}数据，请刷新后重试");
